requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
appnope==0.1.0            # via ipython
asgiref==3.2.9            # via -r requirements/base.txt, django
attrs==19.3.0             # via pytest
backcall==0.2.0           # via ipython
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
coverage==5.2             # via pytest-cov
decorator==4.4.2          # via ipython, traitlets
django-annoying==0.10.6   # via -r requirements/base.txt, django-viewflow
django-debug-toolbar==2.2  # via -r requirements/dev.in
django-environ==0.4.5     # via -r requirements/dev.in
django-filter==2.3.0      # via -r requirements/base.txt, django-viewflow
django-jsonstore==0.4.1   # via -r requirements/base.txt, django-viewflow
django-material==1.6.3    # via -r requirements/base.txt
django-viewflow==1.6.0    # via -r requirements/base.txt
django==3.0.7             # via -r requirements/base.txt, django-annoying, django-debug-toolbar, django-filter, django-jsonstore, django-viewflow, djangorestframework
djangorestframework==3.11.0  # via -r requirements/base.txt
factory-boy==2.12.0       # via -r requirements/dev.in
faker==4.1.1              # via factory-boy
flake8==3.8.3             # via -r requirements/dev.in
httpretty==1.0.2          # via -r requirements/dev.in
idna==2.10                # via requests
ipdb==0.13.3              # via -r requirements/dev.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.15.0           # via ipdb
jedi==0.17.1              # via ipython
mccabe==0.6.1             # via flake8
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest, pytest-sugar
parso==0.7.0              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.5     # via ipython
psycopg2==2.8.5           # via -r requirements/base.txt
ptyprocess==0.6.0         # via pexpect
py==1.8.2                 # via pytest
pycodestyle==2.6.0        # via flake8
pyflakes==2.2.0           # via flake8
pygments==2.6.1           # via ipython
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.0        # via -r requirements/dev.in
pytest-django==3.9.0      # via -r requirements/dev.in
pytest-dotenv==0.5.2      # via -r requirements/dev.in
pytest-sugar==0.9.3       # via -r requirements/dev.in
pytest==5.4.3             # via -r requirements/dev.in, pytest-cov, pytest-django, pytest-dotenv, pytest-sugar
python-dateutil==2.8.1    # via faker
python-dotenv==0.13.0     # via pytest-dotenv
pytz==2020.1              # via -r requirements/base.txt, django
requests==2.24.0          # via -r requirements/dev.in
six==1.15.0               # via -r requirements/base.txt, django-annoying, django-jsonstore, django-material, django-viewflow, packaging, python-dateutil, traitlets
sqlparse==0.3.1           # via -r requirements/base.txt, django, django-debug-toolbar
termcolor==1.1.0          # via pytest-sugar
text-unidecode==1.3       # via faker
traitlets==4.3.3          # via ipython
urllib3==1.25.9           # via requests
wcwidth==0.2.4            # via prompt-toolkit, pytest

# The following packages are considered to be unsafe in a requirements file:
# setuptools
